# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "上傳影片練習用",
        "title": "上傳影片",
        "termsOfService": "https://google.com",
        "contact": {
            "name": "Neil_Hsieh",
            "url": "https://google.com",
            "email": "neil605164@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9999",
    "basePath": "/",
    "paths": {
        "/backend/upload_video": {
            "post": {
                "description": "上傳 mp4 影片",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "上傳影片",
                "parameters": [
                    {
                        "type": "file",
                        "description": "影片檔",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/structs.APIResult"
                        }
                    },
                    "400": {
                        "description": "異常錯誤",
                        "schema": {
                            "$ref": "#/definitions/structs.APIResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "structs.APIResult": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer"
                },
                "error_msg": {
                    "type": "string"
                },
                "log_id": {
                    "type": "string"
                },
                "result": {
                    "type": "object"
                }
            }
        }
    }
}